Pull metadata from databases, do recursive merge
The live config plugin now looks at each database for
the presence of a __metadata table. If found, it is parsed
and included in the metadata.

The __metadata table schema expected:

        CREATE TABLE [__metadata] (
           [key] TEXT PRIMARY KEY,
           [value] TEXT
        );

Where key is a key on the database config for the table. Value
will be JSON parsed and included in the database's metadata.